Added Authentication to Journal Endpoints

+++ src/main/java/net/journal/JournalApp/service/JournalEntryService.java
@@ -23,12 +23,10 @@ public class JournalEntryService {
     @Transactional//to achieve atomicity..everything is considered 1 op...if anything fails...everything rolled back.
     public void saveJournalEntry(JournalEntry journalEntry, String username){
         User user = userService.findByUsername(username);
-        journalEntry.setDate(LocalDateTime.now());
-
-        JournalEntry saved = journalEntryRepo.save(journalEntry);
-        user.getJournalEntries().add(saved);
+        user.getJournalEntries().add(journalEntry);
+        journalEntryRepo.save(journalEntry);
 //        user.setUsername(null);//if this is null then below will fail but still journal entry will be added.
-        userService.saveEntry(user);
+        userService.saveExistingEntry(user);
 
     }
 
@@ -39,11 +37,24 @@ public class JournalEntryService {
     public List<JournalEntry> getAll(){
         return journalEntryRepo.findAll();
     }
-    public void deleteJournalEntry(ObjectId myID, String username){
-        User user = userService.findByUsername(username);
-        user.getJournalEntries().removeIf(x->x.getId().equals(myID));//done becoz in user db, ref of deleted ob is still present. to remove that, we need to do cascade delete
-        userService.saveEntry(user);
-        journalEntryRepo.deleteById(myID);
+
+    @Transactional
+    public boolean deleteJournalEntry(ObjectId myID, String username){
+        boolean removed = false;
+        try{
+            User user = userService.findByUsername(username);
+            removed=user.getJournalEntries().removeIf(x->x.getId().equals(myID));//done becoz in user db, ref of deleted ob is still present. to remove that, we need to do cascade delete
+            if(removed)
+            {
+                userService.saveExistingEntry(user);
+                journalEntryRepo.deleteById(myID);
+            }
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+        return removed;
+
+
     }
     public Optional<JournalEntry> findByID(ObjectId myID){
         return  journalEntryRepo.findById(myID);
